+++ hospitalApi/src/main/java/com/hospitalApi/consults/dtos/CreateConsultRequestDTO.java
@@ -1,5 +1,7 @@
 package com.hospitalApi.consults.dtos;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.NotBlank;
@@ -19,7 +21,7 @@ public class CreateConsultRequestDTO {
     @NotNull(message = "El costo de la consulta es requerido")
     @DecimalMin(value = "0.01", inclusive = true, message = "El costo de la consulta debe ser mayor a 0")
     @Digits(integer = 10, fraction = 2, message = "El costo de la consulta debe tener un máximo de 10 dígitos enteros y 2 decimales")
-    Double costoConsulta;
+    BigDecimal costoConsulta;
 
     @NotBlank(message = "EL id del empleado es requerido")
     String employeeId;
