+++ hospitalApi/src/main/java/com/hospitalApi/consults/dtos/UpdateConsultRequestDTO.java
@@ -1,5 +1,7 @@
 package com.hospitalApi.consults.dtos;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 import jakarta.validation.constraints.DecimalMin;
@@ -18,5 +20,5 @@ public class UpdateConsultRequestDTO {
 
     @DecimalMin(value = "0.01", inclusive = true, message = "El costo de la consulta debe ser mayor a 0")
     @Digits(integer = 10, fraction = 2, message = "El costo de la consulta debe tener un máximo de 10 dígitos enteros y 2 decimales")
-    Double costoConsulta;
+    BigDecimal costoConsulta;
 }
